refactor types and fix tests
