+++ src/types.ts
@@ -1,5 +1,7 @@
 import { ReplaceOperation, AddOperation } from 'fast-json-patch';
 
+type ChangeType = 'breaking' | 'non-breaking' | 'unclassified';
+
 export interface DiffOutput {
   action: string;
   path: string;
@@ -7,5 +9,13 @@ export interface DiffOutput {
   before?: any;
   after?: any;
 }
+
+export type OverrideObject = {
+  [key: string]: {
+    add: ChangeType;
+    remove: ChangeType;
+    edit: ChangeType;
+  };
+};
 
 export type ValueOperation = ReplaceOperation<any> | AddOperation<any>;
