+++ test/fixtures/mergeStandard.fixture.ts
